requirements.txt
numpy==1.12.1
scipy==0.19.0
pillow==4.0.0
